+++ cli/src/program.rs
@@ -35,7 +35,7 @@ use {
     solana_client::{
         connection_cache::ConnectionCache,
         send_and_confirm_transactions_in_parallel::{
-            send_and_confirm_transactions_in_parallel_blocking, SendAndConfirmConfig,
+            send_and_confirm_transactions_in_parallel_blocking_v2, SendAndConfirmConfigV2,
         },
         tpu_client::{TpuClient, TpuClientConfig},
     },
@@ -47,7 +47,7 @@ use {
     solana_rpc_client::rpc_client::RpcClient,
     solana_rpc_client_api::{
         client_error::ErrorKind as ClientErrorKind,
-        config::{RpcAccountInfoConfig, RpcProgramAccountsConfig, RpcSendTransactionConfig},
+        config::{RpcAccountInfoConfig, RpcProgramAccountsConfig},
         filter::{Memcmp, RpcFilterType},
         request::MAX_MULTIPLE_ACCOUNTS,
     },
@@ -1575,7 +1575,11 @@ fn process_program_upgrade(
         let signers = &[fee_payer_signer, upgrade_authority_signer];
         tx.try_sign(signers, blockhash)?;
         let final_tx_sig = rpc_client
-            .send_and_confirm_transaction_with_spinner(&tx)
+            .send_and_confirm_transaction_with_spinner_and_config(
+                &tx,
+                config.commitment,
+                config.send_transaction_config,
+            )
             .map_err(|e| format!("Upgrading program failed: {e}"))?;
         let program_id = CliProgramId {
             program_id: program_id.to_string(),
@@ -1728,10 +1732,7 @@ fn process_set_authority(
             .send_and_confirm_transaction_with_spinner_and_config(
                 &tx,
                 config.commitment,
-                RpcSendTransactionConfig {
-                    preflight_commitment: Some(config.commitment.commitment),
-                    ..RpcSendTransactionConfig::default()
-                },
+                config.send_transaction_config,
             )
             .map_err(|e| format!("Setting authority failed: {e}"))?;
 
@@ -1790,10 +1791,7 @@ fn process_set_authority_checked(
             .send_and_confirm_transaction_with_spinner_and_config(
                 &tx,
                 config.commitment,
-                RpcSendTransactionConfig {
-                    preflight_commitment: Some(config.commitment.commitment),
-                    ..RpcSendTransactionConfig::default()
-                },
+                config.send_transaction_config,
             )
             .map_err(|e| format!("Setting authority failed: {e}"))?;
 
@@ -2132,10 +2130,7 @@ fn close(
     let result = rpc_client.send_and_confirm_transaction_with_spinner_and_config(
         &tx,
         config.commitment,
-        RpcSendTransactionConfig {
-            preflight_commitment: Some(config.commitment.commitment),
-            ..RpcSendTransactionConfig::default()
-        },
+        config.send_transaction_config,
     );
     if let Err(err) = result {
         if let ClientErrorKind::TransactionError(TransactionError::InstructionError(
@@ -2358,10 +2353,7 @@ fn process_extend_program(
     let result = rpc_client.send_and_confirm_transaction_with_spinner_and_config(
         &tx,
         config.commitment,
-        RpcSendTransactionConfig {
-            preflight_commitment: Some(config.commitment.commitment),
-            ..RpcSendTransactionConfig::default()
-        },
+        config.send_transaction_config,
     );
     if let Err(err) = result {
         if let ClientErrorKind::TransactionError(TransactionError::InstructionError(
@@ -2936,7 +2928,11 @@ fn send_deploy_messages(
             } else {
                 initial_transaction.try_sign(&[fee_payer_signer], blockhash)?;
             }
-            let result = rpc_client.send_and_confirm_transaction_with_spinner(&initial_transaction);
+            let result = rpc_client.send_and_confirm_transaction_with_spinner_and_config(
+                &initial_transaction,
+                config.commitment,
+                config.send_transaction_config,
+            );
             log_instruction_custom_error::<SystemError>(result, config)
                 .map_err(|err| format!("Account allocation failed: {err}"))?;
         } else {
@@ -3003,14 +2999,15 @@ fn send_deploy_messages(
                         .expect("Should return a valid tpu client")
                     );
 
-                    send_and_confirm_transactions_in_parallel_blocking(
+                    send_and_confirm_transactions_in_parallel_blocking_v2(
                         rpc_client.clone(),
                         tpu_client,
                         &write_messages,
                         &[fee_payer_signer, write_signer],
-                        SendAndConfirmConfig {
+                        SendAndConfirmConfigV2 {
                             resign_txs_count: Some(max_sign_attempts),
                             with_spinner: true,
+                            rpc_send_transaction_config: config.send_transaction_config,
                         },
                     )
                 },
@@ -3046,10 +3043,7 @@ fn send_deploy_messages(
                     .send_and_confirm_transaction_with_spinner_and_config(
                         &final_tx,
                         config.commitment,
-                        RpcSendTransactionConfig {
-                            preflight_commitment: Some(config.commitment.commitment),
-                            ..RpcSendTransactionConfig::default()
-                        },
+                        config.send_transaction_config,
                     )
                     .map_err(|e| format!("Deploying program failed: {e}"))?,
             ));