+++ client-test/tests/send_and_confirm_transactions_in_parallel.rs
@@ -1,8 +1,9 @@
 use {
     solana_client::{
         nonblocking::tpu_client::TpuClient,
+        rpc_config::RpcSendTransactionConfig,
         send_and_confirm_transactions_in_parallel::{
-            send_and_confirm_transactions_in_parallel_blocking, SendAndConfirmConfig,
+            send_and_confirm_transactions_in_parallel_blocking_v2, SendAndConfirmConfigV2,
         },
     },
     solana_rpc_client::rpc_client::RpcClient,
@@ -51,14 +52,21 @@ fn test_send_and_confirm_transactions_in_parallel_without_tpu_client() {
     let original_alice_balance = rpc_client.get_balance(&alice.pubkey()).unwrap();
     let (messages, sum) = create_messages(alice_pubkey, bob_pubkey);
 
-    let txs_errors = send_and_confirm_transactions_in_parallel_blocking(
+    let txs_errors = send_and_confirm_transactions_in_parallel_blocking_v2(
         rpc_client.clone(),
         None,
         &messages,
         &[&alice],
-        SendAndConfirmConfig {
+        SendAndConfirmConfigV2 {
             with_spinner: false,
             resign_txs_count: Some(5),
+            rpc_send_transaction_config: RpcSendTransactionConfig {
+                skip_preflight: false,
+                preflight_commitment: Some(CommitmentConfig::confirmed().commitment),
+                encoding: None,
+                max_retries: None,
+                min_context_slot: None,
+            },
         },
     );
     assert!(txs_errors.is_ok());
@@ -109,14 +117,21 @@ fn test_send_and_confirm_transactions_in_parallel_with_tpu_client() {
     );
     let tpu_client = rpc_client.runtime().block_on(tpu_client_fut).unwrap();
 
-    let txs_errors = send_and_confirm_transactions_in_parallel_blocking(
+    let txs_errors = send_and_confirm_transactions_in_parallel_blocking_v2(
         rpc_client.clone(),
         Some(tpu_client),
         &messages,
         &[&alice],
-        SendAndConfirmConfig {
+        SendAndConfirmConfigV2 {
             with_spinner: false,
             resign_txs_count: Some(5),
+            rpc_send_transaction_config: RpcSendTransactionConfig {
+                skip_preflight: false,
+                preflight_commitment: Some(CommitmentConfig::confirmed().commitment),
+                encoding: None,
+                max_retries: None,
+                min_context_slot: None,
+            },
         },
     );
     assert!(txs_errors.is_ok());